RPM for zorba and mp-xquery
Zorba is not our own and we make a special install in /opt/zorba.
Note that compiling Zorba require 2.8 or later of cmake which do
not exist on CentOS 6. For this reason the build process
relies on a recent cmake in /opt/cmake . For now, zorba is just
one big package - both runtime and development combined.
mp-xquery is same skeleton as cf-engine and relies on zorba
to be installed.